backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.3
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.1.0
tabulate==0.9.0
pydantic==2.5.0
aiofiles==23.2.1
chromadb==0.4.24
sentence-transformers==2.7.0
transformers==4.40.0
torch>=2.0.0
safetensors>=0.4.0
